Kalman little revisions: linearization removed, doubles replaced by floats; calibration routine in proccess; found a bug with inclination angles non-linearity

+++ SW/Chrono-Terminal-V1.0/firmware/app/inc/filter.h
@@ -66,10 +66,10 @@ typedef struct{
     int16_t     preVal;
     int16_t     V;
     int16_t     a;
-    double      F;
-    double      gain;
-    double      covariance;
-    double      P;
+    float       F;
+    float       gain;
+    float       covariance;
+    float       P;
 }kAxis_type;
 
 typedef struct{
